+++ src/multiRunner.ts
@@ -1,47 +1,21 @@
 import * as vscode from 'vscode';
-import { escapeShell, getEnvCommands } from './util';
-import { RunnerConfig } from './runnerConfig';
-import { TestCase } from './testCase';
-
-interface RunCommand {
-  terminal:string,
-  commands: string[];
-}
-interface DebugCommand {
-  documentUri: vscode.Uri;
-  config: vscode.DebugConfiguration;
-}
+import { TestCase, RunCommand, DebugCommand } from './testCase';
 
 export class MultiRunner {
   private previousRunCommand: RunCommand | undefined;
   private previousDebugCommand: DebugCommand | undefined;
 
-  constructor() {
-  }
-
   public async runTest(testcase:TestCase, options?: string[]): Promise<void> {
-    const config = new RunnerConfig(testcase.filePath);
     const cmds = [];
-    // cwd
-    RunnerConfig.changeDirectoryToWorkspaceRoot && cmds.push(`cd ${escapeShell(config.projectPath)}`);
-    // setting envs
-    const envs = getEnvCommands(config.playwrightEnvironmentVariables);
 
-    cmds.push(...envs);
-    cmds.push(testcase.buildRunCommand(options));
+    const cmd = testcase.buildRunCommand(options);
 
-    await this.executeRunCommand({
-      terminal: 'playwright',
-      commands: cmds,
-    });
+    await this.executeRunCommand(cmd);
   }
 
   public async debugTest(testcase:TestCase, options?: unknown): Promise<void> {
-    let debugConfig = testcase.buildDebugCommand(options);
-    await this.executeDebugCommand({
-      config: debugConfig,
-      documentUri: testcase.filePath,
-    });
+    const cmd = testcase.buildDebugCommand(options);
+    await this.executeDebugCommand(cmd);
   }
   
   public async runPreviousTest(): Promise<void> {
@@ -65,7 +39,7 @@ export class MultiRunner {
     this.previousRunCommand = cmd;
     this.previousDebugCommand = undefined;
 
-    await this.runTerminalCommand(cmd.terminal, ...cmd.commands);
+    await MultiRunner.startTerminal(cmd.options, cmd.command);
   }
 
   private async executeDebugCommand(cmd: DebugCommand) {
@@ -75,10 +49,10 @@ export class MultiRunner {
     await vscode.debug.startDebugging(vscode.workspace.getWorkspaceFolder(cmd.documentUri), cmd.config);
   }
 
-  public async runTerminalCommand(terminalName:string, ...commands: string[]) {
-    let terminal = vscode.window.terminals.find(terminal => terminal.name === terminalName);
+  private static async startTerminal(options:vscode.TerminalOptions, ...commands: string[]) {
+    let terminal = vscode.window.terminals.find(terminal => terminal.name === options.name);
     if (!terminal) {
-      terminal = vscode.window.createTerminal(terminalName);
+      terminal = vscode.window.createTerminal(options);
     }
     terminal.show();
     commands.forEach( command => terminal?.sendText(command) );