+++ src/util.ts
@@ -1,44 +1,10 @@
 import * as path from 'path';
 import * as fs from 'fs';
 import * as vscode from 'vscode';
+import * as child_process from "child_process";
 
 export function isWindows(): boolean {
   return process.platform.includes('win32');
-}
-
-function getShellType(): string {
-  if(isWindows()) {
-    const shell:any = vscode.workspace.getConfiguration().get('terminal.integrated.shell');
-    if(shell.windows){
-      return shell.windows.split('.')[0];
-    }
-    const defaultProfile:any = vscode.workspace.getConfiguration().get('terminal.integrated.defaultProfile');
-    if(defaultProfile.windows){
-      switch(defaultProfile.windows){
-        case 'PowerShell':return 'powershell';
-        case 'Command Prompt':return 'cmd';
-        default:return 'sh';
-      }
-    }
-    return 'powershell';
-  }
-  return 'sh';
-}
-
-export function getEnvCommands(envs:string[]): string[] {
-  const shell = getShellType();
-  return envs.map(env => {
-    const [key,val] = env.split('=');
-    return getCommandToSetEnv(shell, key, val);
-  });
-}
-
-function getCommandToSetEnv(shell:string, key:string, val:string): string {
-  switch (shell) {
-      case 'powershell':return `$ENV:${key}="${val}"`;
-      case 'cmd':return `set ${key}=${val}`;
-  }
-  return `export ${key}=${val}`;
 }
 
 export function normalizePath(path: string): string {
@@ -184,4 +150,13 @@ export class PredefinedVars {
   public set(key:string, val:string){
       this.replaceMap.set("${"+key+"}", val);
   }
+}
+  
+export async function executeBackend(cmd:string): Promise<void> {
+  vscode.window.showInformationMessage(cmd);
+  child_process.exec(cmd, (error, stdout, stderr) => {
+    if(error instanceof Error) {
+      vscode.window.showErrorMessage(error.message);
+    } 
+  });
 }